# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POS Kafe Multi-Cabang API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: tableId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/items:
    post:
      operationId: addOrderItem
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
  /orders/{id}/items/{itemId}:
    patch:
      operationId: updateOrderItem
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
    delete:
      operationId: removeOrderItem
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /orders/{id}/void:
    post:
      operationId: voidOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoidInput"
      responses:
        "200":
          description: Order voided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # KITCHEN DISPLAY
...
  /payments:
    post:
      operationId: processPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    get:
      operationId: getPayments
      tags: [payments]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /payments/{id}/refund:
    post:
      operationId: refundPayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundInput"
      responses:
        "200":
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # STOCK / INGREDIENTS
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, tableId, tableNumber, status, branchId, items, subtotal, tax, total, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        tableId:
          type: integer
        tableNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, served, completed, voided]
        branchId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discountAmount:
          type: number
        tax:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [tableId, branchId, items]
      properties:
        tableId:
          type: integer
        branchId:
          type: integer
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        notes:
          type: ["string", "null"]
        discountAmount:
          type: number
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, served, completed, voided]
        notes:
          type: ["string", "null"]
        discountAmount:
          type: number
    OrderItem:
      type: object
      required: [id, orderId, menuItemId, menuItemName, quantity, unitPrice, totalPrice, kitchenStatus, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        menuItemImage:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        kitchenStatus:
          type: string
          enum: [new, processing, ready, served]
        station:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
    OrderItemUpdate:
      type: object
      properties:
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
    VoidInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        authorizerId:
          type: ["integer", "null"]

    # KITCHEN
...
    Payment:
      type: object
      required: [id, orderId, amount, method, status, branchId, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        orderNumber:
          type: ["string", "null"]
        amount:
          type: number
        method:
          type: string
          enum: [cash, card, qris, ewallet, transfer]
        status:
          type: string
          enum: [pending, completed, refunded, failed]
        change:
          type: ["number", "null"]
        referenceNumber:
          type: ["string", "null"]
        branchId:
          type: integer
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [orderId, amount, method, branchId]
      properties:
        orderId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, card, qris, ewallet, transfer]
        cashReceived:
          type: ["number", "null"]
        referenceNumber:
          type: ["string", "null"]
        branchId:
          type: integer
    RefundInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        amount:
          type: ["number", "null"]

    # INGREDIENT / STOCK
...