# lib/api-zod/src/generated/types/voidInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * POS Kafe Multi-Cabang API
 * OpenAPI spec version: 0.1.0
 */

export interface VoidInput {
  reason: string;
  /** @nullable */
  authorizerId?: number | null;
}